Updated Library
Restructured library to eliminate global variables and to reduce memory footprint.  
Updated documentation for functions
Removed Broken GetCurrent() function
Added working GetCurrentA() and GetCurrentB()
Removed the EnableA/B functions, replacing them with ToggleA/B, to allow for the motors to be dynamically toggled on/off with a single function call
Added constraints on the PWM limits of 98% to avoid Mosfet overheating/overloading
Added warning Pragma to the braking function due to it's testing being limited.  Needs more tests to ensure that it doesn't damage driver
Added function to return current values if pins are assigned outside of default pins